+++ src/test/kotlin/me/bossm0n5t3r/leetcode/missingNumber/MissingNumberTest.kt
@@ -1,10 +1,10 @@
-package missingNumber
+package me.bossm0n5t3r.leetcode.missingNumber
 
 import org.junit.jupiter.api.Test
 import kotlin.test.assertEquals
 
 class MissingNumberTest {
-    private val missingNumber = MissingNumber()
+    private val missingNumber = MissingNumber.Solution()
 
     data class MissingNumberTestData(
         val nums: IntArray,
